basic-llm-demo/pyproject.toml
[project]
name = "basic-llm-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
